pyproject.toml
[project]
name = "OCR-Translate-Sketch"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "mouse>=0.7.1",
    "pandas>=2.3.2",
    "pillow>=11.3.0",
    "pynput>=1.8.1",
    "pystray>=0.19.5",
    "pytesseract>=0.3.13",
    "pyyaml>=6.0.2",
    "watchdog>=6.0.0",
    "win10toast-persist>=0.10.1",
]

[tool.ruff]
line-length = 120
target-version = "py311"

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # Pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
    "S",  # flake8-bandit
    "RUF", # ruff
    "SIM", # flake8-simplify
    "RET", # flake8-return
]

fixable = [
    "I",  # isort
    "W292", # No newline at end of file
    "RUF022", # `__all__` is not sorted
]

[tool.ruff.lint.per-file-ignores]
"tests/*.py" = ["S101"] # Use of assert detected.

[dependency-groups]
dev = [
    "pytest>=8.4.2",
    "pytest-cov>=7.0.0",
]
